requirements.txt
Django>=1.8,<1.9
argparse==1.2.1
docutils==0.12
gunicorn==19.3.0
wsgiref==0.1.2
lockfile==0.10.2
python-daemon==2.0.5
requests==2.6.2
pycrypto==2.6.1
pip-tools==0.3.6

# This won't install on anything but a pi
# so you may need to comment it out.
wiringpi2==1.1.1